feat: restructure API files
- Moved API-related modules (`analyze.ts`, `beans.ts`, `brews.ts`) to a dedicated `/api` directory for better organization.
- Updated import paths in relevant files to reflect the new structure.
